# Conflicts:
#	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/xml/StaxXmlParser.scala
#	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/xml/XmlInferSchema.scala
#	sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/xml/XmlSuite.scala

+++ mllib/src/main/scala/org/apache/spark/mllib/fpm/AssociationRules.scala
@@ -16,7 +16,7 @@
  */
 package org.apache.spark.mllib.fpm
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 import scala.reflect.ClassTag
 
 import org.apache.spark.annotation.Since
@@ -91,8 +91,8 @@ class AssociationRules private[fpm] (
       .map { case (antecedent, ((consequent, freqUnion), freqAntecedent)) =>
         new Rule(antecedent.toArray,
           consequent.toArray,
-          freqUnion,
-          freqAntecedent,
+          freqUnion.toDouble,
+          freqAntecedent.toDouble,
           // the consequent contains always only one element
           itemSupport.get(consequent.head))
       }.filter(_.confidence >= minConfidence)
